// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-audit-entry-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminAuditEntryHttpResponse {
    action?: string;
    actorUserId?: string;
    createdAtUtc?: string;
    id?: string;
    metadata?: string | null;
    subjectClientUserId?: string | null;
    targetId?: string | null;
    targetType?: string;
}