// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-audit-page-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminAuditEntryHttpResponse } from './admin-audit-entry-http-response';

export interface AdminAuditPageHttpResponse {
    items?: Array<AdminAuditEntryHttpResponse>;
    totalItems?: number;
}